pyproject.toml
[tool.poetry]
name = "Fast-Tortoise"
version = "0.1.0"
description = "Fastapi with tortoise and pytest"
authors = ["<yfnel@yfnel.pl>"]
license = "Proprietary"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.135.1"}
pydantic = "^2.12.5"
pydantic-settings = "^2.13.1"
tortoise-orm = "^1.1.7"

[tool.poetry.extras]
utils = ["ipython", "nest-asyncio"]

[tool.poetry.group.dev.dependencies]
pytest = "^9.0.2"
pytest-cov = "^7.0.0"
pytest-asyncio = "^1.3.0"
pytest-socket = "^0.7.0"
ruff = "^0.15.6"

[tool.poetry.group.utils]
optional = true

[tool.poetry.group.utils.dependencies]
ipython = "^9.11.0"
nest-asyncio = "^1.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "--disable-socket --allow-unix-socket"

[tool.tortoise]
tortoise_orm = "config.TORTOISE_ORM"

[tool.ruff]
line-length = 120
target-version = "py313"

[tool.ruff.lint]
select = ["ALL"]
ignore = ["D1", "D203", "D213", "FIX002"]
exclude = ["*migrations*", "local/*"]

[tool.ruff.lint.per-file-ignores]
"test_*.py" = ["ANN001", "ANN201", "FBT003", "PLR0913", "PLR2004", "S101", "S105"]

[tool.ruff.lint.flake8-quotes]
docstring-quotes = "double"
inline-quotes = "single"

